Move CBlockUndo from main to undo, update CTxUndo
vprevout in CTxUndo is a vector of Coins now, serialization is left to TxInUndoSerializer and TxInUndoDeserializer,
which receive pointer to the Coin in the ctor. Coinbase and coinstake flags, height and version fields are removed
because Coin already has them. Type and version is removed from Serialize and Unserialize functions.
CBlockUndo::WriteToDisk and ReadFromDisk are moved to src/validation. They are now global functions, which receive
CBlockUndo from the outside